déplacement des methodes isFirstSlotOfBooking et formatBookingTime dans le bookingStore
